+++ Lab1-1_new/Lab1-1_new/RandFile.cpp
@@ -3,7 +3,7 @@
 void RandFile(const char* filename,int size)
 {
 	srand(time(0)); //избавл€емс€ от псевдорандома прив€зкой ко времени
-	int* rand_array = new int[size]; //выдел€ем пам€ть под массив
+	double* rand_array = new double[size]; //выдел€ем пам€ть под массив
 
 	for (int i=0;i<size;i++) //заполн€ем массив рандомными значени€ми
 	{
